Split FTs into separate files
